Controllers are done. Now I only need to cleanup my views, set up routing and do other small tasks like these.

+++ ERP_testTask/ERP_testTask/Controllers/MoviesController.cs
@@ -7,38 +7,48 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using ERP_testTask.Models;
+using Microsoft.AspNet.Identity;
+using PagedList;
 
 namespace ERP_testTask.Controllers
 {
+    [Authorize]
     public class MoviesController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: Movies
-        public ActionResult Index()
+        [AllowAnonymous]
+        public ActionResult Index(int? page)
         {
-            return View(db.Movies.ToList());
+            int pageSize = 10;
+            int pageNumber = (page ?? 1);
+            return View(db.Movies.OrderByDescending(movie => movie.Id).ToPagedList(pageNumber, pageSize));
         }
 
         // GET: Movies/Details/5
+        [AllowAnonymous]
         public ActionResult Details(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return HttpNotFound();
             }
+
             Movie movie = db.Movies.Find(id);
+
             if (movie == null)
             {
                 return HttpNotFound();
             }
+
             return View(movie);
         }
 
         // GET: Movies/Create
         public ActionResult Create()
         {
-            return View();
+            return View(new MovieCreateModel());
         }
 
         // POST: Movies/Create
@@ -46,16 +56,32 @@ namespace ERP_testTask.Controllers
         // сведения см. в статье https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Name,Description,Year,Director,PosterURL")] Movie movie)
+        public ActionResult Create(MovieCreateModel model)
         {
             if (ModelState.IsValid)
             {
+                var movie = new Movie
+                {
+                    Name = model.Name,
+                    Description = model.Description,
+                    Year = model.Year,
+                    Director = model.Director,
+                    UserId = User.Identity.GetUserId()
+                };
+
+                if (model.Poster != null)
+                {
+                    string fileName = System.IO.Path.GetFileName(model.Poster.FileName);
+                    model.Poster.SaveAs(Server.MapPath("~/Posters/" + model.Name + "_" + fileName));
+                    movie.PosterURL = "/Posters/" + model.Name + "_" + fileName;
+                }
+
                 db.Movies.Add(movie);
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Details", new { id = movie.Id });
             }
 
-            return View(movie);
+            return View(model);
         }
 
         // GET: Movies/Edit/5
@@ -63,14 +89,25 @@ namespace ERP_testTask.Controllers
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return HttpNotFound();
             }
+
             Movie movie = db.Movies.Find(id);
+
             if (movie == null)
             {
                 return HttpNotFound();
             }
-            return View(movie);
+
+            var model = new MovieCreateModel
+            {
+                Name = movie.Name,
+                Description = movie.Description,
+                Year = movie.Year,
+                Director = movie.Director
+            };
+
+            return View(model);
         }
 
         // POST: Movies/Edit/5
@@ -78,28 +115,42 @@ namespace ERP_testTask.Controllers
         // сведения см. в статье https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Name,Description,Year,Director,PosterURL")] Movie movie)
-        {
-            if (ModelState.IsValid)
-            {
-                db.Entry(movie).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            return View(movie);
-        }
-
-        // GET: Movies/Delete/5
-        public ActionResult Delete(int? id)
+        public ActionResult Edit(Int32? id, MovieCreateModel model)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return HttpNotFound();
             }
+
             Movie movie = db.Movies.Find(id);
+
             if (movie == null)
             {
                 return HttpNotFound();
+            }
+
+            if (movie.UserId != User.Identity.GetUserId())
+            {
+                return HttpNotFound();
+            }
+
+            if (ModelState.IsValid)
+            {
+                movie.Name = model.Name;
+                movie.Description = model.Description;
+                movie.Year = model.Year;
+                movie.Director = model.Director;
+
+                if (model.Poster != null)
+                {
+                    string fileName = System.IO.Path.GetFileName(model.Poster.FileName);
+                    model.Poster.SaveAs(Server.MapPath("~/Posters/" + model.Name + "_" + fileName));
+                    movie.PosterURL = "/Posters/" + model.Name + "_" + fileName;
+                }
+
+                db.Entry(movie).State = EntityState.Modified;
+                db.SaveChanges();
+                return RedirectToAction("Details", new { id = movie.Id });
             }
             return View(movie);
         }
@@ -107,21 +158,28 @@ namespace ERP_testTask.Controllers
         // POST: Movies/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int? id)
         {
+            if (id == null)
+            {
+                return HttpNotFound();
+            }
+
             Movie movie = db.Movies.Find(id);
+
+            if (movie == null)
+            {
+                return HttpNotFound();
+            }
+
+            if (movie.UserId != User.Identity.GetUserId())
+            {
+                return HttpNotFound();
+            }
+
             db.Movies.Remove(movie);
             db.SaveChanges();
             return RedirectToAction("Index");
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
         }
     }
 }
